Fix nasty function overloading bug in pinhole_K()
The `pinhole_K()` function was overloaded with:

- fx, fy, cx, cy
- intrinsics vector
- image_width, image_height, lens_hfov, lens_vfov

The problem with the last overload is in terms of data type it looks
similar to the first overload (fx, fy, cx, cy). The difference is only
between an `int` and `double`. The image width and height are normally
stored in a `vec2_t` and therefore the compiler would choose the first
overload over the third.

As such to mitigate this confusion the third overload has been
refactored to:

- image_size, lens_hfov, lens_vfov

Or:

    mat3_t pinhole_K(const vec2_t &image_size,
                     const double lens_hfov,
                     const double lens_vfov);